feat: add image upload pipeline, storage backends, albums, and image API
Phase 3: core image hosting capability.

- Storage interface with local filesystem and S3-compatible backends
- Upload service implementing the 13-step flow (auth, validation, rate limit,
  image processing, dedup, storage write, DB record, thumbnail generation)
- Image processor: format conversion, quality re-encoding via imaging library
- Thumbnail generator (400px max, best-effort async)
- Path/filename naming rules with token expansion ({Y}/{m}/{d}/{uniqid}/etc)
- Image API: upload (multipart), list, get, delete, update
- Album API: full CRUD with ownership validation
- Rate limiting per time window (minute/hour/day/month)
- MD5+SHA1 deduplication scoped per storage strategy

+++ internal/sqlc/queries/users.sql
@@ -42,7 +42,7 @@ UPDATE users SET image_num = image_num + 1, updated_at = NOW() WHERE id = $1;
 UPDATE users SET image_num = image_num - 1, updated_at = NOW() WHERE id = $1 AND image_num > 0;
 
 -- name: GetUserUsedCapacity :one
-SELECT COALESCE(SUM(size_bytes), 0) FROM images WHERE user_id = $1;
+SELECT COALESCE(SUM(size_bytes), 0)::bigint FROM images WHERE user_id = $1;
 
 -- name: CreateAdminUser :one
 INSERT INTO users (group_id, email, password, name, role, capacity_bytes, settings, status, email_verified, registered_ip)
